requirements.txt
nsetools
numpy
pymongo
requests
beautifulsoup4
lxml
tqdm

python-dotenv==0.14.0
sqlalchemy~=2.0.12
psycopg2-binary
alembic~=1.10.4
envyaml~=1.10.211231

apscheduler==3.10.4

# FastAPI dependencies
fastapi==0.61.1
uvicorn==0.12.1
pydantic==1.10.7
python-jose[cryptography]
passlib[bcrypt]